+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Added debug log emits for the ``DataFrameETL`` transformer (#24, #27)
+
 ## [0.1.6] - 2018-1-12
 
 ### Fixed